Moved custom models to their own models directory and file. Updated imports accordingly.
