+++ mvvm/src/models/blog/blogGenreModel.ts
@@ -1,0 +1,10 @@
+const genreOptions = [
+  "educational",
+  "excursion",
+  "review",
+  "comparison",
+  "tutorial",
+  "news",
+] as const;
+
+export type BlogGenre = (typeof genreOptions)[number];
