+++ mvvm/src/models/blog/blogPostImageModel.ts
@@ -1,0 +1,10 @@
+export type Image = {
+  id: number;
+  status: string;
+};
+
+export interface BlogPostImage {
+  blogPostId: string;
+  id: string;
+  imageId: string;
+}
