+++ mvvm/src/models/blog/newBlogModel.ts
@@ -1,0 +1,7 @@
+export type NewBlog = {
+  title: string;
+  description: string;
+  content: string;
+  featured: string;
+  images?: string[];
+};
