+++ mvvm/src/models/likeModel.ts
@@ -1,0 +1,6 @@
+export type Like = {
+  id: string;
+  commentId: string;
+  userId: string;
+  createdAt: string;
+};
