+++ mvvm/src/models/messageModel.ts
@@ -1,0 +1,9 @@
+export type Message = {
+  id: string;
+  subject: string;
+  message: string;
+  name: string;
+  email: string;
+  createdAt: string;
+  read: boolean;
+};
