+++ mvvm/src/models/termModel.ts
@@ -1,0 +1,4 @@
+export type Term = {
+  word: string;
+  definition: string;
+};
